drivers/clk: fix conflicting types in clk_register_* definitions
Update the function definitions in all 6 clk implementation files to
match the uintptr_t parameter type already declared in clk_provider.h.

Fixes CI error:
  error: conflicting types for 'clk_register_divider'

+++ drivers/clk/clk_mux.c
@@ -177,7 +177,7 @@ const struct clk_ops_s g_clk_mux_ro_ops =
 FAR struct clk_s *clk_register_mux(FAR const char *name,
                                    FAR const char * const *parent_names,
                                    uint8_t num_parents,
-                                   uint8_t flags, uint32_t reg,
+                                   uint8_t flags, uintptr_t reg,
                                    uint8_t shift, uint8_t width,
                                    uint8_t clk_mux_flags)
 {
